requirements.txt
anyio==3.5.0
asgiref==3.5.0
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.75.2
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
orjson==3.6.8
pathspec==0.9.0
platformdirs==2.5.2
pydantic==1.9.0
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
tomli==2.0.1
typing-extensions==4.2.0
ujson==5.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
